Combat and Exp
Added lines for giving out exp at end of battle, so characters will now
level up(forgot earlier). Also fixed some combat buggyness.

+++ src/Party/BadParty.java
@@ -48,7 +48,7 @@ public class BadParty extends Party{
 		//Not currently Implemented
 	}
 	
-	public void addEXP(int EXP){
+	public void addEXP(double EXP){
 		//Not currently Implemented
 	}
