+++ src/Party/BossParty.java
@@ -27,7 +27,7 @@ public class BossParty extends Party{
 		//Not currently Implemented
 	}
 	
-	public void addEXP(int EXP){
+	public void addEXP(double EXP){
 		//Not currently Implemented
 	}
